WIP - Refactor appliances, separate from results
