+++ frontend/app/components/appliances-forms/sharedForms.module.scss
@@ -1,0 +1,88 @@
+// .container {
+//   display: flex;
+//   flex-direction: column;
+//   align-items: center;
+//   padding: 20px;
+// }
+
+.dropdown {
+  width: auto;
+  padding: 5px;
+  margin-bottom: 20px;
+  margin-top: 1rem; // distance form label
+}
+
+.form {
+  display: flex;
+  flex-direction: column;
+  row-gap: 0.5rem;
+  width: 100%;
+  margin-bottom: 1rem;
+}
+
+.sharedInputGroup {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 1rem;
+}
+
+.radioGroup {
+  @extend .sharedInputGroup;
+
+  // .radioLabel {
+  //   display: flex;
+  //   align-items: center;
+  //   row-gap: 10px;
+  // }
+
+  .radioInput {
+    margin-right: 2px;
+  }
+
+  .radioText {
+    padding-left: 2px;
+  }
+
+  .radioOptions {
+    display: flex;
+    flex-direction: column;
+    row-gap: 10px;
+
+    @media (min-width: 768px) {
+      flex-direction: row;
+      column-gap: 10px;
+    }
+  }
+}
+
+.sliderGroup {
+  @extend .sharedInputGroup;
+  input[type="range"] {
+    width: 300px;
+  }
+}
+
+.submitButton {
+  display: inline-block;
+  padding: 10px 20px;
+  background-color: rgb(249, 214, 91);
+  color: black;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+
+  &:hover {
+    background-color: rgb(213, 185, 84);
+  }
+}
+
+.labelWithInfo {
+  display: flex;
+  align-items: center;
+  margin-bottom: 0.4rem;
+}
+
+// @media (min-width: 768px) {
+//   .container {
+//   }
+// }
